Add files via upload
Initial commit (if I understand the term correctly). So far, this repository has the capability to pull data from every submission in reddit.com/r/rational (or any other targeted subreddit) all the way back to early 2013. I'm not quite sure yet why it goes back only that far-- it could be 1.) The subreddit submissions only go back that far (even though this particular subreddit was founded in 2009) or 2.) The API I'm using is hard-limited in that regard, perhaps by Reddit itself.